float to double

+++ Coursework1/MultipleLinearRegresion.hpp
@@ -4,8 +4,8 @@ class MultipleLinearRegresion : public RegressionBase
 {
 
 public:
-    vector<double> element_wise_multiplication(vector<double> x1, vector<double> x2) {
-        vector<double> vec;
+    vector<float> element_wise_multiplication(vector<float> x1, vector<float> x2) {
+        vector<float> vec;
 
         for (int i = 0; x1.size() < n; i++) {
             vec.push_back(x1[i]*x2[i]);
@@ -13,36 +13,36 @@ public:
         return(vec);
     }
 
-    double sum(vector<double> vec) {
-        double summation = 0;
+    float sum(vector<float> vec) {
+        float summation = 0;
         for (int i = 0; vec.size() < n; i++) {
             summation += vec[i];
         }
         return(summation);
     }
 
-    void fit(vector<double> x1, vector<double> x2, vector<double> y, int solver, bool verbose=false)
+    void fit(vector<float> x1, vector<float> x2, vector<float> y, int solver, bool verbose=false)
     {
-        vector<vector<double>> mat = dummy_mat;
-        vector<double> b = dummy_vec;
+        vector<vector<float>> mat = dummy_mat;
+        vector<float> b = dummy_vec;
         n = x1.size();
 
         // row1
-        vector<double> row1;
+        vector<float> row1;
         row1.push_back(n);
         row1.push_back(sum(x1));
         row1.push_back(sum(x2));
         mat.push_back(row1);
 
         // row2
-        vector<double> row2;
+        vector<float> row2;
         row2.push_back(sum(x1));
         row2.push_back(sum(element_wise_multiplication(x1, x1)));
         row2.push_back(sum(element_wise_multiplication(x1, x2)));
         mat.push_back(row2);
 
         // row3
-        vector<double> row3;
+        vector<float> row3;
         row3.push_back(sum(x2));
         row3.push_back(sum(element_wise_multiplication(x1, x2)));
         row3.push_back(sum(element_wise_multiplication(x2, x2)));
